// src/Store/Ducks/Users/types.ts
export enum UserActionTypes {
  POST_USER_REQUEST = 'POST_USER_REQUEST',
  POST_USER_SUCCESS = 'POST_USER_SUCCESS',
  POST_USER_FAILURE = 'POST_USER_FAILURE',
  POST_LOGIN_REQUEST = 'POST_LOGIN_REQUEST',
  POST_LOGIN_SUCCESS = 'POST_LOGIN_SUCCESS',
  POST_LOGIN_FAILURE = 'POST_LOGIN_FAILURE',
  DELETE_USER_REQUEST = 'DELETE_USER_REQUEST',
  DELETE_USER_SUCCESS = 'DELETE_USER_SUCCESS',
  DELETE_USER_FAILURE = 'DELETE_USER_FAILURE',
  GET_USERS_REQUEST = 'GET_USERS_REQUEST',
  GET_USERS_SUCCESS = 'GET_USERS_SUCCESS',
  GET_USERS_FAILURE = 'GET_USERS_FAILURE',
  USER_LOGOUT = 'USER_LOGOUT'
}